+++ packages/vue/src/sdk.ts
@@ -1,10 +1,10 @@
 import { getDefaultIntegrations, init as browserInit } from '@sentry/browser';
 import type { Client } from '@sentry/core';
 import { applySdkMetadata, setNormalizeStringifier } from '@sentry/core/browser';
-import { normalizeStringifyValue as browserNormalizeStringifyValue } from '@sentry-internal/browser-utils';
+
 import { vueIntegration } from './integration';
-import { isVueViewModel } from './isVueViewModel';
 import type { Options } from './types';
+import { normalizeStringifyValue } from './isVueViewModel';
 
 /**
  * Inits the Vue SDK
@@ -23,11 +23,4 @@ export function init(options: Partial<Omit<Options, 'tracingOptions'>> = {}): Cl
   setNormalizeStringifier(normalizeStringifyValue);
 
   return client;
-}
-
-function normalizeStringifyValue(value: Exclude<unknown, string | number | boolean | null>): string | undefined {
-  if (isVueViewModel(value)) {
-    return (value as { __v_isVNode?: boolean }).__v_isVNode ? '[VueVNode]' : '[VueViewModel]';
-  }
-  return browserNormalizeStringifyValue(value);
 }
